+++ plugins/commands/destroy/command.rb
@@ -25,17 +25,25 @@ module VagrantPlugins
         return if !argv
 
         @logger.debug("'Destroy' each target VM...")
-        declined = false
+        declined = 0
+        total    = 0
         with_target_vms(argv, :reverse => true) do |vm|
           action_env = vm.action(
             :destroy, :force_confirm_destroy => options[:force])
 
-          declined = true if action_env.has_key?(:force_confirm_destroy_result) &&
+          total    += 1
+          declined += 1 if action_env.has_key?(:force_confirm_destroy_result) &&
             action_env[:force_confirm_destroy_result] == false
         end
 
-        # Success if no confirms were declined
-        declined ? 1 : 0
+        # Nothing was declined
+        return 0 if declined == 0
+
+        # Everything was declined
+        return 1 if declined == total
+
+        # Some was declined
+        return 2
       end
     end
   end