+++ CHANGELOG.md
@@ -24,6 +24,8 @@ FEATURES:
 
 IMPROVEMENTS:
 
+  - commands/destroy: Exit codes changes. 0 means everything succeeded.
+      1 means everything was declined. 2 means some were declined. [GH-811]
   - commands/ssh-config: Works without a target in multi-machine envs [GH-2844]
   - guests/freebsd: Support for virtio interfaces. [GH-3082]
   - guests/openbsd: Support for virtio interfaces. [GH-3082]